setFeeSplit(): test for arithmetic overflow if payDebt parameter is larger than total debt

+++ solidity/contracts/interfaces/IPCV.sol
@@ -34,7 +34,7 @@ interface IPCV {
 
     function setFeeRecipient(address _feeRecipient) external;
 
-    function setFeeSplit(uint256 _feeSplitPercentage) external;
+    function setFeeSplit(uint8 _feeSplitPercentage) external;
 
     function withdrawMUSD(address _recipient, uint256 _musdAmount) external;
